block delete for received_flag = 1, remove reply_to mail new user

+++ app/Http/Controllers/Admin/DeliveryCrudController.php
@@ -748,32 +748,36 @@ class DeliveryCrudController extends CrudController
     {
         $delivery = Delivery::where('id', $id)->first();
         if (isset($delivery)) {
-            // IssuedMaterialOuthouse::where('ds_num', $delivery->ds_num)
-            //                     ->where('ds_line', $delivery->ds_line)
-            //                     ->delete();
-            $imos = IssuedMaterialOuthouse::where('ds_num', $delivery->ds_num)
-            ->where('ds_line', $delivery->ds_line)->get();
-            if($imos->count() > 0){
-                foreach($imos as $imo){
-                    $imo->delete();
-                }
-            }
-            $ds = Delivery::where('id', $id)->get();
-            if($ds->count() > 0){
-                foreach($ds as $d){
-                    $d->delete();
-                }
-            }
-            // Delivery::where('id', $id)->delete();
-            // DeliveryStatus::where('ds_num', $delivery->ds_num)
-            //                 ->where('ds_line', $delivery->ds_line)
-            //                 ->delete();
+            $receivedFlag = 0;
+            
+
             $dds = DeliveryStatus::where('ds_num', $delivery->ds_num)
             ->where('ds_line', $delivery->ds_line)->get();
             if($dds->count() > 0){
                 foreach($dds as $dd){
-                    $dd->delete();
+                    $receivedFlag += $dd->received_flag;
+                    if ($dd->received_flag == 0) {
+                        $dd->delete();
+                    }
                 }
+            }
+
+            if ($receivedFlag == 0) {
+                $imos = IssuedMaterialOuthouse::where('ds_num', $delivery->ds_num)
+                ->where('ds_line', $delivery->ds_line)->get();
+                if($imos->count() > 0){
+                    foreach($imos as $imo){
+                        $imo->delete();
+                    }
+                }
+                $ds = Delivery::where('id', $id)->get();
+                if($ds->count() > 0){
+                    foreach($ds as $d){
+                        $d->delete();
+                    }
+                }
+            }else{
+                return response()->json(['message' => "Failure! Data has been received"], 403);
             }
         }
